+++ api/src/models/index.ts
@@ -1,0 +1,38 @@
+export { default as AlerteMessageModel } from "./alerteMessage";
+export { default as ApplicationModel } from "./application";
+export { default as AreasModel } from "./areas";
+export { default as BusModel } from "./bus";
+export { default as CohesionCenterModel } from "./cohesionCenter";
+export { default as CohortModel } from "./cohort";
+export { default as ContractModel } from "./contract";
+export { default as DepartmentServiceModel } from "./departmentService";
+export { default as EmailModel } from "./email";
+export { default as EventModel } from "./event";
+export { default as FiltersModel } from "./filters";
+export { default as InscriptionGoalModel } from "./inscriptionGoal";
+export { default as MeetingPointModel } from "./meetingPoint";
+export { default as MissionModel } from "./mission";
+export { default as MissionAPIModel } from "./missionAPI";
+export { default as MissionEquivalenceModel } from "./missionEquivalence";
+export { default as ProgramModel } from "./program";
+export { default as ReferentModel } from "./referent";
+export { default as SchoolModel } from "./school";
+export { default as SchoolRAMSESModel } from "./schoolRAMSES";
+export { default as SessionPhase1Model } from "./sessionPhase1";
+export { default as SessionPhase1TokenModel } from "./sessionPhase1Token";
+export { default as StructureModel } from "./structure";
+export { default as TagsModel } from "./tags";
+export { default as WaitingListModel } from "./waitingList";
+export { default as YoungModel } from "./young";
+
+export { default as CleClasseModel } from "./cle/classe";
+export { default as CleEtablissement } from "./cle/etablissement";
+
+export { default as ImportPlanTransportModel } from "./PlanDeTransport/importPlanTransport";
+export { default as LigneBusModel } from "./PlanDeTransport/ligneBus";
+export { default as LigneToPointModel } from "./PlanDeTransport/ligneToPoint";
+export { default as ModificationBusModel } from "./PlanDeTransport/modificationBus";
+export { default as PlanTransportModel } from "./PlanDeTransport/planTransport";
+export { default as PointDeRassemblementModel } from "./PlanDeTransport/pointDeRassemblement";
+export { default as SchemaDeRepartitionModel } from "./PlanDeTransport/schemaDeRepartition";
+export { default as TableDeRepartitionModel } from "./PlanDeTransport/tableDeRepartition";
