+++ packages/lib/src/constants/plan-de-transport.js
@@ -25,7 +25,7 @@ const PDT_IMPORT_ERRORS_TRANSLATION = {
   BAD_CLASSE_ID: { text: "ID de classe %s non existant.", tooltip: "Les ID des classes sont transmis dans le schéma de répartition." },
   BAD_PDR_ID: { text: "ID du PDR %s non existant.", tooltip: "Les ID des PDR sont transmis dans le schéma de répartition." },
   BAD_PDR_DEPARTEMENT: { text: "Erreur sur le département du PDR %s.", tooltip: "Le département du fichier ne correspond pas au département du PDR en base de données." },
-  BAD_MERGED_LINE_ID: { text: "Numéro de ligne fusionnée %s inconnu", tooltip: "Les lignes fusionnées doivent être présentent dans le fichier." },
+  BAD_MERGED_LINE_ID: { text: "Numéro de ligne fusionnée %s inconnu", tooltip: "Les lignes fusionnées doivent être présentes dans le fichier." },
   SAME_PDR_ON_LINE: { text: "Un même PDR %s apparaît plusieurs fois sur la ligne.", tooltip: "Un PDR ne peut appraître qu'une seule fois sur une même ligne." },
   LACKING_YOUNG_CAPACITY: {
     text: "La somme de la capacité volontaires des lignes est inférieure à la somme des volontaires du schéma de répartition affecté dans le centre",
